Cooldown visuals and logic from PR #2032 only: Cooldown.tsx, KamiCard.tsx integration, calc base fallback
