+++ packages/client/src/app/components/library/cards/KamiCard/KamiCard.tsx
@@ -9,14 +9,9 @@ import { Kami } from 'network/shapes/Kami';
 import { getItemImage } from 'network/shapes/utils/images';
 import { playClick } from 'utils/sounds';
 import { Card } from '../';
-import { Cooldown } from './Cooldown';
+import { Cooldown, useCooldownVisuals } from './Cooldown';
 import { Health } from './Health';
-import { calcCooldown, calcCooldownRequirement, onCooldown } from 'app/cache/kami/calcs/base';
-import { SteamShader } from 'app/components/shaders/SteamShader';
-import { LightningShader } from 'app/components/shaders/LightningShader';
-import { ShaderStack } from 'app/components/shaders/ShaderStack';
-import { makeStaticLayer } from 'app/components/shaders/StaticShader';
-import { CRTShader } from 'app/components/shaders/CRTShader';
+import { onCooldown } from 'app/cache/kami/calcs/base';
 
 // KamiCard is a card that displays information about a Kami. It is designed to display
 // information ranging from current harvest or death as well as support common actions.
@@ -105,30 +100,10 @@ export const KamiCard = ({
     }));
   }, [getTempBonuses, kami]);
 
-  // compute cooldown progress
-  const totalCooldown = calcCooldownRequirement(kami);
-  const remaining = calcCooldown(kami);
-  const progress = totalCooldown > 0 ? Math.min(1, Math.max(0, remaining / totalCooldown)) : 0;
-  const shaped = Math.pow(progress, 1.25); // slightly eased fade
-
-  // grayscale amount should go from 1 -> 0 over cooldown
-  const grayAmount = shaped; // 1 at start, 0 at end
-
-  // Drive periodic re-render while on cooldown so CSS filter progresses
-  const isOnCooldown = onCooldown(kami);
-  const [cooldownTick, setCooldownTick] = useState(0);
-  useEffect(() => {
-    if (!isOnCooldown) return;
-    const id = setInterval(() => setCooldownTick((t) => (t + 1) % 1000000), 200);
-    return () => clearInterval(id);
-  }, [isOnCooldown]);
-
-  // Final-second uniform static wipe (wait 0.5s, then 0.5s wipe)
-  const lastSecond = remaining <= 1.0;
-  const timeIntoLast = lastSecond ? Math.max(0, 1.0 - remaining) : 0;
-  const wipeWait = 0.5;
-  const wipeDur = 0.5;
-  const wipeProgress = lastSecond ? Math.max(0, Math.min(1, (timeIntoLast - wipeWait) / wipeDur)) : 0;
+  const { filter: cooldownFilter, foreground: cooldownForeground } = useCooldownVisuals(
+    kami,
+    !!showCooldown,
+  );
 
   const TitleSection = (
     <TitleBar>
@@ -149,63 +124,9 @@ export const KamiCard = ({
         showLevelUp: showLevelUp && canLevel,
         showSkillPoints: showSkillPoints && (kami.skills?.points ?? 0) > 0,
         onClick: handleKamiClick,
-        filter:
-          showCooldown && onCooldown(kami)
-            ? (grayAmount > 0 ? `grayscale(${grayAmount}) contrast(1.05)` : undefined)
-            : undefined,
+        filter: cooldownFilter,
         background: undefined,
-        foreground:
-          showCooldown && onCooldown(kami)
-            ? (() => {
-                // Static grain and wipe are updated per-frame via onBeforeFrame
-                const staticLayer: any = makeStaticLayer({ brightness: 1.6, alpha: 0.96, vertical: true });
-                staticLayer.onBeforeFrame = (uniforms: any) => {
-                  const tot = calcCooldownRequirement(kami);
-                  const rem = calcCooldown(kami);
-                  const prog = tot > 0 ? Math.min(1, Math.max(0, rem / tot)) : 0;
-                  const eased = Math.pow(prog, 1.25);
-                  const alpha = 0.96 * eased;
-                  if (uniforms.uAlpha) uniforms.uAlpha.value = alpha;
-                };
-
-                // Final-second uniform static wipe: wait 0.5s, then 0.5s fade
-                const wipeLayer: any = makeStaticLayer({
-                  brightness: 1.7,
-                  alpha: 0.0,
-                  vertical: true,
-                  topFeather: 0.0,
-                  maskRadius: 0.0,
-                  maskHeight: 0.0,
-                });
-                wipeLayer.onBeforeFrame = (uniforms: any) => {
-                  const tot = calcCooldownRequirement(kami);
-                  const rem = calcCooldown(kami);
-                  const lastSecond = rem <= 1.0 && rem > 0; // only show while > 0 and <= 1s
-                  if (!lastSecond) {
-                    if (uniforms.uAlpha) uniforms.uAlpha.value = 0.0;
-                    return;
-                  }
-                  const timeIntoLast = Math.max(0, 1.0 - rem);
-                  const wait = 0.5;
-                  const dur = 0.5;
-                  const wp = Math.max(0, Math.min(1, (timeIntoLast - wait) / dur));
-                  const a = 0.9 * (1 - wp);
-                  if (uniforms.uTopSplit) uniforms.uTopSplit.value = 2.0; // uniform
-                  if (uniforms.uAlpha) uniforms.uAlpha.value = a;
-                };
-
-                return (
-                  <>
-                    {/* Subtle CRT layer remains constant */}
-                    <CRTShader brightness={1.6} alpha={0.96} />
-                    {/* Static grain fades over full cooldown */}
-                    <ShaderStack layers={[staticLayer]} animateWhenOffscreen />
-                    {/* Final-second wipe */}
-                    <ShaderStack layers={[wipeLayer]} animateWhenOffscreen />
-                  </>
-                );
-              })()
-            : undefined,
+        foreground: cooldownForeground,
       }}
     >
       {TitleSection}
